move component header into same folderr
