+++ demo/components/GridShowcase.vue
@@ -144,7 +144,7 @@
 </template>
 
 <script>
-import ComponentHeader from '../ComponentHeader'
+import ComponentHeader from './ComponentHeader'
 
 export default {
   name: 'grid-showcase',
